+++ test/git_hub_actions/versions_test.exs
@@ -19,7 +19,7 @@ defmodule GitHubActions.VersionsTest do
 
   test "from_config/0" do
     assert List.last(Versions.from_config()) ==
-             [otp: ["27.0/2"], elixir: ["1.17.0/3", "1.18.0"]]
+             [otp: ["27.0/2"], elixir: ["1.17.0/3", "1.18.0/1"]]
   end
 
   describe "get/2" do
@@ -135,8 +135,9 @@ defmodule GitHubActions.VersionsTest do
                "1.17.1",
                "1.17.2",
                "1.17.3",
-               # v1.18.0
-               "1.18.0"
+               # v1.18.0/1
+               "1.18.0",
+               "1.18.1"
              ]
     end
 
@@ -290,7 +291,8 @@ defmodule GitHubActions.VersionsTest do
                %Version{major: 1, minor: 17, patch: 1},
                %Version{major: 1, minor: 17, patch: 2},
                %Version{major: 1, minor: 17, patch: 3},
-               %Version{major: 1, minor: 18, patch: 0}
+               %Version{major: 1, minor: 18, patch: 0},
+               %Version{major: 1, minor: 18, patch: 1}
              ]
     end
 
@@ -328,7 +330,7 @@ defmodule GitHubActions.VersionsTest do
                %Version{major: 1, minor: 15, patch: 8},
                %Version{major: 1, minor: 16, patch: 3},
                %Version{major: 1, minor: 17, patch: 3},
-               %Version{major: 1, minor: 18, patch: 0}
+               %Version{major: 1, minor: 18, patch: 1}
              ]
     end
   end
@@ -459,7 +461,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 22, minor: 3},
-                 elixir: %Version{major: 1, minor: 18, patch: 0}
+                 elixir: %Version{major: 1, minor: 18, patch: 1}
                ],
                [
                  otp: %Version{major: 23, minor: 3},
@@ -475,7 +477,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 23, minor: 3},
-                 elixir: %Version{major: 1, minor: 18, patch: 0}
+                 elixir: %Version{major: 1, minor: 18, patch: 1}
                ],
                [
                  otp: %Version{major: 24, minor: 3},
@@ -483,7 +485,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 24, minor: 3},
-                 elixir: %Version{major: 1, minor: 18, patch: 0}
+                 elixir: %Version{major: 1, minor: 18, patch: 1}
                ],
                [
                  otp: %Version{major: 25, minor: 3},